[level 0] Title: 배열의 평균값, Time: 0.03 ms, Memory: 73 MB -BaekjoonHub

+++ 프로그래머스/0/120817. 배열의 평균값/배열의 평균값.java
@@ -1,14 +1,14 @@
 class Solution {
     public double solution(int[] numbers) {
         double answer = 0;
-        double sum = 0;
+        int sum = 0;
         int len = numbers.length;
         
         for(int i=0;i<len;i++){
             sum += numbers[i];
         }
         
-        answer = sum / len;
+        answer = (double)sum / len;
         return answer;
     }
 }
